Restart application if was destroyed and opened from recents fixes https://github.com/remotv/controller-for-android/issues/114
